requirements.txt
streamlit>=1.32
google-genai>=0.3
anthropic>=0.40
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
youtube-transcript-api>=0.6
yt-dlp>=2024.1
requests>=2.31
python-dotenv>=1.0
